Common BSD: Add --gc-sections ld flags
As common bsd's are using the -ffunction-sections flags from
mkspecs/common/gcc-base.conf, also use the --gc-sections ld flags
like Linux as all are using GNU ld as their linker and all are
capable of using it. The last remaining problems with --gc-sections
removing the .name sections providing the OS tags for distinction of
the execution layer were solved with NetBSD 6.0 and OpenBSD 5.4, see
http://gnats.netbsd.org/40401 and http://www.openbsd.org/plus54.html
so this option can safely be used now to reduce the size of moc and rcc
tools.

Change-Id: I74ccd4f6bc607f6d82d32fc864875f26b26bf167
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>