Revert "End the drag if the dockwidget is being hidden"
This reverts commit 75b705fec8e9517047d7dfa98203edff69f2bf8a
and 798e0064e9be78f8320ff25a9af50d1b5e5badb1.

This broke drag and drop of QDockWidget if there is no animations
(QTBUG-54185)  and this caused numerous crash with GrouppedDragging:
crashes when dropping on another floating QDockWidget, or when
dragging away.

The problem being that 'endDrag' would cause the QDockWidget to be hiden
temporarily and it would recurse into 'endDrag' and have unwanted effects

Task-number: QTBUG-54185
Change-Id: Icad9c6748448783e039bde205b60133f89f25e8e
Reviewed-by: Andy Shaw <andy.shaw@qt.io>