Include Qt 5.6 change to 5.6.1  https://codereview.qt-project.org/#/c/159690/
QNSView: Reset the internal buttons state after dragging ends

We only track left and right mouse buttons when dragging. But
some applications may do this with other mouse buttons.
In this case, instead of tracking which button was pressed and
which one was released, we just ask Cocoa for the current state.

Task-number: QTBUG-53374