Include Qt 5.6 change to 5.6.1   https://codereview.qt-project.org/#/c/153420/
Cocoa QPA: Mute mouse dragging warnings

While probably valid, these warnings are beyond the
user's control, and more likely a sign of Qt's own
expectations or misbehavior. So, we should not annoy
the users with them.
This change introduces the same logging category as
3ee01f74031d9e1d0 in 5.7 as is a partial backport to
simplify a subsequent 5.6 to 5.7 merge.

Task-number: QTBUG-42846