Limit the number of targets the VNC platform plugin is be built for
The VNC platform plugin is a port from the 4.8 code, which only
supported embedded Linux.  Right now that is the only platform this is
tested with so unless there is a need for other platforms to be hosting
applications over VNC we should limit it to only being built on unix
platforms that are not Android, or Darwin.

Change-Id: I49407669ee8ca6e9e3ed95b5bb350c7730839cf8
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>