include QtC WIP patch   https://codereview.qt-project.org/#/c/157023/
WIP Cocoa: Workaround rdar #25880259

We found out that it's not possible to delete an NSView
without its containing NSWindow keeping a dangling reference
to it. It's not an issue when this happens as a consequence
of a mouse up event, but it's a different story if it happens
from a mouse down or mouse dragged events.

What we do is that we set a flag in -[QNSView mouseDown:] which
get cleared in -[QNSView mouseUp:]. If, when we reach ~QCocoaWindow,
the flag is set on its QNSView, then we remove the QNSView form the
views hierarchy and cache it (instead of just releasing it). When
-[NSWindow sendEvent:] finally receives the mouse up event, we clear
the cache.